+++ src/dataflow.rs
@@ -2,9 +2,33 @@
 //! sequences.
 
 use crate::Constrain;
-use crate::DataFlow;
 use crate::Iterable;
 use crate::Sequence;
+
+pub trait DataFlow<T> {
+    //! A trait for very local dataflow. It's generic across `T`, a type intended to represent
+    //! "things" a machine instruction may read from or write to.
+    //!
+    //! For example, a type representing a Z80 machine code instruction could implement this for
+    //! the Z80's register file, the flags, the I/O space and the address space.
+
+    /// returns true iff the variable `t` is read (used) by the instruction or basic block before
+    /// any assignment. Such a variables must be live at the start of the block.
+    fn reads(&self, t: &T) -> bool;
+
+    /// returns true iff the variable `t` is assigned (written to) by the instruction or basic
+    /// block, effectively "killing" any previous value it held.
+    fn writes(&self, t: &T) -> bool;
+
+    /// Modifies the instruction
+    fn modify(&mut self) -> bool;
+
+    /// Modifies the instruction so that it reads from `t`.
+    fn make_read(&mut self, t: &T) -> bool;
+
+    /// Modifies the instruction so that it writes to `t`.
+    fn make_write(&mut self, t: &T) -> bool;
+}
 
 /// A constraint for asserting that a datum (that is, the type parameter representing, say, a
 /// register, status flag, memory location, etc.) is not live at the beginning of a sequence.
