+++ src/lib.rs
@@ -102,16 +102,6 @@ pub trait Constrain<Insn> {
     }
 }
 
-pub trait CallingConvention<SamplePoint, InputParameters, ReturnValue> {
-    //! A trait for calling conventions. A type which implements this trait can execute a function
-    //! taking the given argument(s), and return the function's return value.
-
-    /// Calls the given callable object, passing it the parameters of type `InputParameters`, and returning an
-    /// `ReturnValue`.
-    fn call(function: &SamplePoint, parameters: InputParameters)
-        -> Result<ReturnValue, StropError>;
-}
-
 /// Enumerates reasons why executing a function may fail
 #[derive(Debug, PartialEq)]
 pub enum StropError {
@@ -138,47 +128,5 @@ impl<InputParameters, ReturnValue> Callable<InputParameters, ReturnValue>
 {
     fn call(&self, parameters: InputParameters) -> Result<ReturnValue, StropError> {
         (self)(parameters)
-    }
-}
-
-pub trait DataFlow<T> {
-    //! A trait for very local dataflow. It's generic across `T`, a type intended to represent
-    //! "things" a machine instruction may read from or write to.
-    //!
-    //! For example, a type representing a Z80 machine code instruction could implement this for
-    //! the Z80's register file, the flags, the I/O space and the address space.
-
-    /// returns true iff the variable `t` is read (used) by the instruction or basic block before
-    /// any assignment. Such a variables must be live at the start of the block.
-    fn reads(&self, t: &T) -> bool;
-
-    /// returns true iff the variable `t` is assigned (written to) by the instruction or basic
-    /// block, effectively "killing" any previous value it held.
-    fn writes(&self, t: &T) -> bool;
-
-    /// Modifies the instruction
-    fn modify(&mut self) -> bool;
-
-    /// Modifies the instruction so that it reads from `t`.
-    fn make_read(&mut self, t: &T) -> bool;
-
-    /// Modifies the instruction so that it writes to `t`.
-    fn make_write(&mut self, t: &T) -> bool;
-}
-
-pub trait Peephole {
-    //! A trait for very local peephole optimizations. It's generic across `T`, a type intended to
-    //! represent machine instructions.
-    //!
-    //! The default implementation is effectively a no-op.
-
-    /// Modifies the instruction
-    fn modify(&mut self) -> bool {
-        unreachable!();
-    }
-
-    /// Checks if two instructions may not follow eachother.
-    fn check(_a: &Self, _b: &Self) -> bool {
-        false
     }
 }
